adb: only send close packet in on_child_exit method
There is a situation similar to "adb shell ps", after the device
response data to the host, the host sends okay packet, and then
the device should return close packet. In the original code, it
is possible that the child process is slow to execute, resulting
in the main process receiving okay packet, but the svc has not
been released yet, resulting in no close packet being returned.

Signed-off-by: yangsong8 <yangsong8@xiaomi.com>